Move CLI nonce account helpers in client
