+++ cli/src/nonce.rs
@@ -11,19 +11,13 @@ use clap::{App, Arg, ArgMatches, SubCommand};
 use solana_clap_utils::{
     input_parsers::*, input_validators::*, offline::BLOCKHASH_ARG, ArgConstant,
 };
-use solana_client::{nonce_utils::Error, rpc_client::RpcClient};
+use solana_client::{nonce_utils::*, rpc_client::RpcClient};
 use solana_remote_wallet::remote_wallet::RemoteWalletManager;
 use solana_sdk::{
     account::Account,
-    account_utils::StateMut,
-    commitment_config::CommitmentConfig,
     hash::Hash,
     message::Message,
-    nonce::{
-        self,
-        state::{Data, Versions},
-        State,
-    },
+    nonce::{self, State},
     pubkey::Pubkey,
     system_instruction::{
         advance_nonce_account, authorize_nonce_account, create_nonce_account,
@@ -198,58 +192,6 @@ impl NonceSubCommands for App<'_, '_> {
                 )
                 .arg(nonce_authority_arg()),
         )
-    }
-}
-
-pub fn get_account(rpc_client: &RpcClient, nonce_pubkey: &Pubkey) -> Result<Account, Error> {
-    get_account_with_commitment(rpc_client, nonce_pubkey, CommitmentConfig::default())
-}
-
-pub fn get_account_with_commitment(
-    rpc_client: &RpcClient,
-    nonce_pubkey: &Pubkey,
-    commitment: CommitmentConfig,
-) -> Result<Account, Error> {
-    rpc_client
-        .get_account_with_commitment(nonce_pubkey, commitment)
-        .map_err(|e| Error::Client(format!("{}", e)))
-        .and_then(|result| {
-            result
-                .value
-                .ok_or_else(|| Error::Client(format!("AccountNotFound: pubkey={}", nonce_pubkey)))
-        })
-        .and_then(|a| match account_identity_ok(&a) {
-            Ok(()) => Ok(a),
-            Err(e) => Err(e),
-        })
-}
-
-pub fn account_identity_ok(account: &Account) -> Result<(), Error> {
-    if account.owner != system_program::id() {
-        Err(Error::InvalidAccountOwner)
-    } else if account.data.is_empty() {
-        Err(Error::UnexpectedDataSize)
-    } else {
-        Ok(())
-    }
-}
-
-pub fn state_from_account(account: &Account) -> Result<State, Error> {
-    account_identity_ok(account)?;
-    StateMut::<Versions>::state(account)
-        .map_err(|_| Error::InvalidAccountData)
-        .map(|v| v.convert_to_current())
-}
-
-pub fn data_from_account(account: &Account) -> Result<Data, Error> {
-    account_identity_ok(account)?;
-    state_from_account(account).and_then(|ref s| data_from_state(s).map(|d| d.clone()))
-}
-
-pub fn data_from_state(state: &State) -> Result<&Data, Error> {
-    match state {
-        State::Uninitialized => Err(Error::InvalidStateForOperation),
-        State::Initialized(data) => Ok(data),
     }
 }
 
@@ -669,9 +611,10 @@ mod tests {
     use crate::cli::{app, parse_command};
     use solana_sdk::{
         account::Account,
+        account_utils::StateMut,
         fee_calculator::FeeCalculator,
         hash::hash,
-        nonce::{self, State},
+        nonce::{self, state::Versions, State},
         signature::{read_keypair_file, write_keypair, Keypair, Signer},
         system_program,
     };
