+++ client/Cargo.toml
@@ -20,6 +20,7 @@ serde = "1.0.112"
 serde_derive = "1.0.103"
 serde_json = "1.0.56"
 solana-account-decoder = { path = "../account-decoder", version = "1.4.0" }
+solana-clap-utils = { path = "../clap-utils", version = "1.4.0" }
 solana-net-utils = { path = "../net-utils", version = "1.4.0" }
 solana-sdk = { path = "../sdk", version = "1.4.0" }
 solana-transaction-status = { path = "../transaction-status", version = "1.4.0" }
